+++ src/core/helpers/form-helpers.ts
@@ -32,9 +32,3 @@ export const errorMessage = (
     }
     return '';
 };
-
-export const resetForm = (formData: Record<string, string>): void => {
-    for (const field in formData) {
-        formData[field] = '';
-    }
-};
